Library search path fixes, library configuration dialog fixes, and code cleaning.
Added application method to fix searching for user libraries.
Fixed documentation search path bug.
Moved auto pan setting from draw frame to draw panel were it is defined.
Some minor device context drawing changes.

+++ include/class_base_screen.h
@@ -212,7 +212,7 @@ public:
      */
     double  GetScalingFactor()
     {
-        return (double) m_ZoomScalar / GetZoom();
+        return (double) m_ZoomScalar / (double) GetZoom();
     }
